PegOutRequest struct (#114)
* Modified PegOut struct to match contract implementation
* Removed fee from struct and its usage

+++ examples/union/setup/accept_pegin_setup.rs
@@ -28,7 +28,7 @@ impl AcceptPegInSetup {
         keyring: &Keyring,
         bitvmx: &BitVMXClient,
         committee_id: Uuid,
-        slot_index: usize,
+        slot_index: u64,
         rootstock_address: String,
         reimbursement_pubkey: PublicKey,
     ) -> Result<()> {
